deps/v8/test/inspector/debugger/pause-on-async-call-set-timeout-expected.txt

Checks Debugger.pauseOnAsynCall with setTimeout.

Running test: testSetTimeout
paused at:
debugger; #setTimeout(() => 1, 0);

paused at:
debugger; setTimeout(() => #1, 0);


Running test: testDebuggerStmtBeforeCallback1
paused at:
debugger; #setTimeout(() => 1, 0);debugger;

paused at:
debugger; setTimeout(() => 1, 0);#debugger;

paused at:
#debugger//should-break-here


Running test: testDebuggerStmtBeforeCallback2
paused at:
setTimeout('debugger//should-break-here', 0);
#setTimeout(() => 1, 0);

paused at:
#debugger//should-break-here


Running test: testSetTimeoutWithoutJS
paused at:
debugger; #setTimeout('}', 0);
setTimeout('var a = 239;', 0);

paused at:
#debugger//should-break-here


Running test: testResume
paused at:
#setTimeout(() => 42, 0)

paused at:
setTimeout(() => #42, 0)
